@@ -1,23 +1,38 @@
-import { Router, Switch, Route } from "react-router-dom";
+import React from "react";
+import { Routes, Route, BrowserRouter, MemoryRouter } from "react-router";
+import { PathDef } from "./routing";
 import { NotFoundComponent } from "./components";
-import { history, PathDef } from "./routing";
+import { PLATFORM } from "./env";
 
 export default function App({
   routes,
+  strict = false,
 }: {
   enUSLocale?: boolean;
   routes: { [name: string]: PathDef<any, any> };
+  strict: boolean;
 }) {
-  return (
-    <Router history={history}>
-      <div className="app-body">
-        <Switch>
-          {Object.entries(routes).map(([key, route]) => (
-            <Route key={key} {...(route as PathDef<any, any>).routeDef} />
-          ))}
-          <Route component={NotFoundComponent} />
-        </Switch>
-      </div>
-    </Router>
+  const child = (
+    <div className="app-body">
+      <Routes>
+        {Object.entries(routes).map(([key, route]) => (
+          <Route key={key} {...(route as PathDef<any, any>).routeDef} />
+        ))}
+        <Route element={<NotFoundComponent />} />
+      </Routes>
+    </div>
   );
+
+  let main = undefined;
+
+  if (PLATFORM === "native") {
+    main = <MemoryRouter>{child}</MemoryRouter>;
+  } else {
+    main = <BrowserRouter>{child}</BrowserRouter>;
+  }
+
+  if (strict) {
+    return <React.StrictMode>{main}</React.StrictMode>;
+  }
+  return main;
 }
