@@ -1,66 +1,32 @@
-import { makeStyles } from "@mui/styles";
-import { routeAPIs } from "../routing";
-import { NoArgsPathDef } from "../routing/route";
+import { routeAPIs, NoArgsPathDef } from "../routing";
+import styles from "./NotFoundComponent.module.css";
+import { useNavigate } from "react-router";
 
-const useStyles = makeStyles({
-  container: {
-    fontFamily:
-      "-apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Helvetica Neue, Arial, Noto Sans, sans-serif, apple color emoji, segoe ui emoji, Segoe UI Symbol, noto color emoji",
-    fontSize: 14,
-    textAlign: "center",
-  },
-  title: {
-    lineHeight: 1.8,
-    fontSize: 24,
-  },
-  subTitle: {
-    lineHeight: 1.6,
-    color: "#00000073",
-  },
-  navigation: {
-    marginTop: 24,
-    "& button:not(:first-child)": {
-      marginLeft: 8,
-    },
-  },
-  btn: {
-    color: "#fff",
-    borderColor: "#1890ff",
-    background: "#1890ff",
-    textShadow: "0 -1px 0 rgb(0 0 0 / 12%)",
-    boxShadow: "0 2px #0000000b",
-    lineHeight: 1.5715,
-    display: "inline-block",
-    fontWeight: 400,
-    border: "1px solid transparent",
-    transition: "all .3s cubic-bezier(.645,.045,.355,1)",
-    userSelect: "none",
-    touchAction: "manipulation",
-    height: 32,
-    padding: "4px 15px",
-    borderRadius: 2,
-    cursor: "pointer",
-  },
-} as const);
-
-const onClickGoHome = new NoArgsPathDef({
+const homePath = new NoArgsPathDef({
   component: () => null,
   pathDef: "/",
-}).path().mouseClickNavigationHandler;
+}).path();
 
 const NotFoundComponent = () => {
-  const classes = useStyles();
+  const navigate = useNavigate();
+
   return (
-    <div className={classes.container}>
-      <div className={classes.title}>404</div>
-      <div className={classes.subTitle}>
+    <div className={styles.container}>
+      <div className={styles.title}>404</div>
+      <div className={styles.subTitle}>
         Sorry, the page you visited does not exist.
       </div>
-      <div className={classes.navigation}>
-        <button className={classes.btn} onClick={routeAPIs.goBack}>
+      <div className={styles.navigation}>
+        <button
+          className={styles.btn}
+          onClick={() => routeAPIs.goBack(navigate)}
+        >
           Back
         </button>
-        <button className={classes.btn} onClick={onClickGoHome}>
+        <button
+          className={styles.btn}
+          onClick={homePath.getMouseClickNavigationHandler(navigate)}
+        >
           Home
         </button>
       </div>
