@@ -1,2 +1,2 @@
-export { default as LoadingComponent } from "./LoadingComponent";
 export { default as NotFoundComponent } from "./NotFoundComponent";
+export { default as LoadingComponent } from "./LoadingComponent";
